* alioth/yoh/master: (28 commits)
  BF: Pion/ff got proper function with a docstring to please tests
  BF:daily_tests -- ln datadb only if not already there
  BF: skip glmnet test if glmnet is NA
  RF: make tests.main disable pymvpa warnings for nosetests as well
  RF: move waveletmapper under nosetests (for SkipTest)
  NF: skeleton for knn_plot example
  BF: misc minor in misc.plot.base
  BF: missing import of skip_if_no_external
  RF: making use of skip_if_no_external -- now we are doomed to switch to nose ;)
  BF: making that SMLR custom_target labile test
  RF: we do not care about rpy any longer
  NF: providing SkipTest and skip_if_no_external helper from testing.tools
  RF: 2nd portion of unittests to adhere new setup -- all nonlabile pass now
  BF: missing members of __all__s
  RF: testing would import numpy as N
  RF: adjusting first part of unittests to use new testing/
  RF: test_svmkernels -- made it use mvpa.testing + few additional checks
  BF: adding forgotten svmkernels test
  RF: adjusting testing so test_{clf,rfe} work
  RF: ENFORCE_STATES_ENABLED -> ENFORCE_CA_ENABLED  since no more .states
  ...

+++ tools/daily_tests.sh
@@ -69,7 +69,7 @@ do_checkout() {
     #fi
     $precmd git checkout $branch
     #provide datadb
-    $precmd ln -s "$datadbdir" .
+    [ -e "$datadbdir" ] || $precmd ln -s "$datadbdir" .
 }
 
 do_build() {
